Moved logics to create chooser intent to a separate file
